+++ lib/simulation/Cattle.h
@@ -11,7 +11,7 @@ class Cattle : public Consumer {
 
  public:
   Cattle(std::string consumerId) : Consumer(consumerId) {
-    int energy = config::EnergyPlan::getEnergyFromWattage(2000);
+    double energy = config::EnergyPlan::getEnergyFromWattage(2000);
 
     addEnergyPlan(new config::EnergyPlanStatic(420, 423, energy));
     addEnergyPlan(new config::EnergyPlanStatic(1110, 1113, energy));
