+++ lib/simulation/Consumer.cpp
@@ -17,15 +17,15 @@ void Consumer::dump(std::ostringstream& out) {
   out << "      Consumer-Id: " << this->id << ", rate: " << getCurrentEnergy() << std::endl;
 }
 
-int Consumer::getCurrentEnergy() throw (exception::EnergyException) {
+double Consumer::getCurrentEnergy() throw (exception::EnergyException) {
   if(energyPlans.empty()) return 0.0;
   std::vector<config::EnergyPlan *>::iterator it;
-  float retVal = 0.0;
+  double retVal = 0.0;
   for(it = energyPlans.begin(); it!=energyPlans.end(); it++) {
     retVal += ((config::EnergyPlan *) *it)->getCurrentEnergy();
   }
   if(retVal < 0) throw exception::EnergyException("Device consumes negative energy: Check you configuration file");
-  return (int) retVal;
+  return retVal;
 }
 
 } /* End of namespace simulation::endpoint::consumer */
